Use off_t instead of long for all variables related to position in file:
...offset, cursor position, etc.

+++ src/editor/edit-impl.h
@@ -183,24 +183,24 @@ void user_menu (WEdit * edit, const char *menu_file, int selected_entry);
 void edit_init_menu (struct WMenuBar *menubar);
 void menu_save_mode_cmd (void);
 gboolean edit_translate_key (WEdit * edit, long x_key, int *cmd, int *ch);
-int edit_get_byte (WEdit * edit, long byte_index);
-int edit_get_utf (WEdit * edit, long byte_index, int *char_width);
-long edit_count_lines (WEdit * edit, long current, long upto);
-long edit_move_forward (WEdit * edit, long current, long lines, long upto);
-long edit_move_forward3 (WEdit * edit, long current, long cols, long upto);
-long edit_move_backward (WEdit * edit, long current, long lines);
+int edit_get_byte (WEdit * edit, off_t byte_index);
+int edit_get_utf (WEdit * edit, off_t byte_index, int *char_width);
+long edit_count_lines (WEdit * edit, off_t current, off_t upto);
+off_t edit_move_forward (WEdit * edit, off_t current, long lines, off_t upto);
+off_t edit_move_forward3 (WEdit * edit, off_t current, long cols, off_t upto);
+off_t edit_move_backward (WEdit * edit, off_t current, long lines);
 void edit_scroll_screen_over_cursor (WEdit * edit);
 void edit_render_keypress (WEdit * edit);
 void edit_scroll_upward (WEdit * edit, long i);
 void edit_scroll_downward (WEdit * edit, long i);
 void edit_scroll_right (WEdit * edit, long i);
 void edit_scroll_left (WEdit * edit, long i);
-void edit_move_up (WEdit * edit,  long i, gboolean do_scroll);
+void edit_move_up (WEdit * edit, long i, gboolean do_scroll);
 void edit_move_down (WEdit * edit, long i, gboolean do_scroll);
-void edit_move_to_prev_col (WEdit * edit, long p);
+void edit_move_to_prev_col (WEdit * edit, off_t p);
 long edit_get_col (WEdit * edit);
-long edit_bol (WEdit * edit, long current);
-long edit_eol (WEdit * edit, long current);
+off_t edit_bol (WEdit * edit, off_t current);
+off_t edit_eol (WEdit * edit, off_t current);
 void edit_update_curs_row (WEdit * edit);
 void edit_update_curs_col (WEdit * edit);
 void edit_find_bracket (WEdit * edit);
@@ -214,12 +214,12 @@ void edit_delete_line (WEdit * edit);
 
 int edit_delete (WEdit * edit, const int byte_delete);
 void edit_insert (WEdit * edit, int c);
-void edit_cursor_move (WEdit * edit, long increment);
+void edit_cursor_move (WEdit * edit, off_t increment);
 void edit_push_undo_action (WEdit * edit, long c, ...);
 void edit_push_redo_action (WEdit * edit, long c, ...);
 void edit_push_key_press (WEdit * edit);
 void edit_insert_ahead (WEdit * edit, int c);
-long edit_write_stream (WEdit * edit, FILE * f);
+off_t edit_write_stream (WEdit * edit, FILE * f);
 char *edit_get_write_filter (const vfs_path_t * write_name_vpath,
                              const vfs_path_t * filename_vpath);
 gboolean edit_save_confirm_cmd (WEdit * edit);
@@ -234,7 +234,7 @@ gboolean edit_load_cmd (WEdit * edit, edit_current_file_t what);
 void edit_mark_cmd (WEdit * edit, gboolean unmark);
 void edit_mark_current_word_cmd (WEdit * edit);
 void edit_mark_current_line_cmd (WEdit * edit);
-void edit_set_markers (WEdit * edit, long m1, long m2, long c1, long c2);
+void edit_set_markers (WEdit * edit, off_t m1, off_t m2, long c1, long c2);
 void edit_push_markers (WEdit * edit);
 void edit_replace_cmd (WEdit * edit, int again);
 void edit_search_cmd (WEdit * edit, gboolean again);
@@ -242,12 +242,14 @@ mc_search_cbret_t edit_search_cmd_callback (const void *user_data, gsize char_of
                                             int *current_char);
 void edit_complete_word_cmd (WEdit * edit);
 void edit_get_match_keyword_cmd (WEdit * edit);
-gboolean edit_save_block (WEdit * edit, const char *filename, long start, long finish);
+
+gboolean edit_save_block (WEdit * edit, const char *filename, off_t start, off_t finish);
 gboolean edit_save_block_cmd (WEdit * edit);
 gboolean edit_insert_file_cmd (WEdit * edit);
 void edit_insert_over (WEdit * edit);
-int edit_insert_column_of_text_from_file (WEdit * edit, int file,
-                                          long *start_pos, long *end_pos, long *col1, long *col2);
+off_t edit_insert_column_of_text_from_file (WEdit * edit, int file,
+                                            off_t * start_pos, off_t * end_pos, long *col1,
+                                            long *col2);
 
 long edit_insert_file (WEdit * edit, const vfs_path_t * filename_vpath);
 gboolean edit_load_back_cmd (WEdit * edit);
@@ -256,7 +258,7 @@ void edit_block_process_cmd (WEdit * edit, int macro_number);
 void edit_refresh_cmd (WEdit * edit);
 void edit_date_cmd (WEdit * edit);
 void edit_goto_cmd (WEdit * edit);
-int eval_marks (WEdit * edit, long *start_mark, long *end_mark);
+int eval_marks (WEdit * edit, off_t * start_mark, off_t * end_mark);
 void edit_status (WEdit * edit);
 void edit_execute_key_command (WEdit * edit, unsigned long command, int char_for_insertion);
 void edit_update_screen (WEdit * edit);
@@ -288,7 +290,7 @@ void edit_set_filename (WEdit * edit, const vfs_path_t * name_vpath);
 
 void edit_load_syntax (WEdit * edit, char ***pnames, const char *type);
 void edit_free_syntax_rules (WEdit * edit);
-void edit_get_syntax_color (WEdit * edit, long byte_index, int *color);
+void edit_get_syntax_color (WEdit * edit, off_t byte_index, int *color);
 
 void book_mark_insert (WEdit * edit, long line, int c);
 int book_mark_query_color (WEdit * edit, long line, int c);
@@ -302,7 +304,7 @@ void book_mark_serialize (WEdit * edit, int color);
 void book_mark_restore (WEdit * edit, int color);
 
 gboolean line_is_blank (WEdit * edit, long line);
-long edit_indent_width (WEdit * edit, long p);
+long edit_indent_width (WEdit * edit, off_t p);
 void edit_insert_indent (WEdit * edit, int indent);
 void edit_options_dialog (WEdit * edit);
 void edit_syntax_dialog (WEdit * edit);
